refactor(utils/misc): move get-text functions into separate file & add docstring for them     for #5187
